pyproject.toml
[tool.poetry]
name = "hikari-discord-bot"
version = "0.1.0"
description = "A discord bot with"
authors = ["CharmfulProgrammer"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
hikari-lightbulb = "^2.2.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jurigged = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
